splittin easy problems from difficult ones
